Add support for building with pyinstaller
This change adds basic support for building a bundle using pyinstaller.

At the moment, I wouldn't use this for distributing the bot yet because it does not include the `import.py`. Once that is inlined, we could do it.

Until then, these changes might be good for testing that everything still works as expected in non-bundled mode.

In order to test the pyinstaller version:
1. Install pyinstaller
2. Run `pyinstaller pokebot.py`
3. Go to `dist/pokebot/` and run `pokebot.exe` (of course, need to add ROMs to the `roms/` directory first)