Improve version/commit hash detection
Currently, the stock `version.py` tries to detect the last commit hash and use that as a version number.

This only worked if the head (i.e. most recent commit) of a branch has been checked out, not for arbitrary commits.

So this fixes that. Also, if that commit happens to be a tagged one, it shows the tag name instead.