Add a Map debug tab
This adds a debug tab showing some information about the current map and tile.

While the tab is selected, this also handles clicks on the emulator's video output to select a _different_ tile whose data will then be displayed inside the tab (as long as that other tab is part of the current map.)

This also fixes a few issues with the `Trainer` class on FR/LG, where location data was not returned correctly.