Change how audio is toggled in unthrottled mode, to hopefully fix Win11 audio issues
Unthrottling the emulator causes the audio stream to be stopped. When throttling again the stream is restarted, but apparently this sometimes fails for some reason on Windows 11.

This PR changes the throttling behaviour so that the audio stream will always be closed and recreated respectively.

Hopefully that fixes the issue. At the very least I did not observe any negative impact when testing, so things should at least not get worse.