Extract Item Storage System data
This adds functions to access the Item Storage System (i.e. the PC storage)

It also updates the old item bag accessing functions to use the `Item` data structure that we were already using for Pokémon data and that contains a lot more information about an item type.

It also adds helper functions to figure out whether there is still space for a given item, which might be useful for Pickup handling.