Update how the player's current map is detected
The previous method relied on checking the player avatar object's metadata, which appears to not always be accurate in that regard.

Checking the save data works better and I think we used to do it this way, but there was probably a reason why it had been changed.

Let's see if this comes back to bite us.