Add a listener system for handling overworld events
This adds a bunch of global listeners, i.e. functions that run every frame and wait for certain events to happen.

The idea is that those listeners can look out for events that are could more or less happen at any time and are not closely tied to the bot mode script.

Listeners I have implemented so far:

- Battle starts
- Spotted by trainer
- PokeNav call started
- Egg is hatching
- Repel expired
- Party Pokemon fainted due to poison
- Whiteout
- Safari Zone steps used up

In order to allow these listeners to actually do something about the events they detected, the bot controllers are now organised as a stack of Generator functions.

A listener (or even a bot mode) has the option to push its own Generator onto this stack. The bot will run the most recently pushed Generator function first, and then continue with the previous one etc. until the stack is empty.

The idea here is that these listeners can deal with events that would interrupt the regular flow (such as a PokeNav call) with their own actions (such as: press B until the call is over) while the regular bot mode is suspended. So from the perspective of the bot mode it looks like nothing ever happened.

This makes bot modes a bit easier to develop because they don't all have to look out for these events.

The way the battle system was implemented effectively already worked like this before, so I've just extended that for more events.

I have added some callback methods to bot modes (see the `on_*` functions in `modules/modes/_interface.py`.) Most events get reported to the bot mode that way in case the mode wants to do somehow handle them.