Improve pathfinding to be faster and work across map boundaries
This adds a new and improved pathfinding function (using the A* algorithm) that allows navigating across different maps.

It only works with maps that are directly connected (i.e. that you can just walk between, as opposed to warping through doors, holes in the ground, cave entrances, ...) but it _does_ allow you to specify a warp as a destination and will try its best to use that warp.

This fixes a couple of minor issues with the old pathing functions, such as being able to handle overlapping paths (bridges that you can walk across but also surf under, or the bicycle overpass on Hoenn's route 110.)

Since the behaviour may be slightly different in places, the old navigation function still exists (albeit renamed.) I'd rather not break 50% of modes again by updating them without sufficient testing, so I'll leave that for future PRs.