Split up `_utils` and add a debug list for currently running utils
This splits up the old `_utils.py` into several files to make things a bit easier to find.

While doing so, I have introduced a `@debug.track` decorator that (if Debug mode is enabled) will log all currently running actions/utils and make the list accessible in the `Emulator` debug tab. This offers a bit more introspection without having to run the bot in a debugger.

It also introduces a 'Debug Values' dict. The idea is that you can just write to that dict while developing, and whatever value it contains is then displayed in the `Emulator` debug tab. Depending on the value, this might be more comfortable to view and use than just printing it to the console.

Example code:

```python
from modules.debug import debug
debug.debug_values["Opponent"] = get_opponent()
```