Rock Smash mode: Use Mach Bike in Safari Zone
This updates the Rock Smash mode to use the new `navigate_to()` function.

It also changes the Safari Zone part of the mode so that it can make use of the Mach Bike for longer travelling distances -- if it is registered to the Select button, it will be used.

This is useful for the path between the entrance and the area with the rocks, as well as for getting to a different map and returning in order to respawn the rocks.